refactor: move checksum from internal/hashutil to utils package
- Moved pkg/internal/hashutil/checksum.go to pkg/utils/checksum.go
- Updated all imports across the codebase
- Removed empty internal/hashutil directory
- All tests pass
